FEATURE: create smaller components for ProductDetail component
